refactor(demo): refactor ChatDialog.vue
