+++ demo/src/App.vue
@@ -7,29 +7,20 @@ import { initMcp } from './mcp';
 
 const { endpointTransport, mcpValidator } = initMcp();
 
-const clientId = ref(endpointTransport.clientId);
 if (process.env.NODE_ENV === 'development') {
-  if (!endpointTransport.clientId) {
-    endpointTransport.clientIdResolved.then((id) => {
-      clientId.value = id;
-      console.log('Client ID:', clientId.value);
-    });
-  } else {
-    console.log('Client ID:', clientId.value);
-  }
+  endpointTransport.clientIdResolved.then((id) => {
+    console.log('Client ID:', id);
+  });
 }
 
 const showChat = ref(true);
+
 </script>
 
 <template>
   <AddUser />
-  <ChatDialog
-    :client-id="clientId"
-    :genCode="mcpValidator.genVerifyCode"
-    :clearCode="mcpValidator.clearVerifyCode"
-    v-model:show="showChat"
-  />
+  <ChatDialog :get-client-id="() => endpointTransport.clientId" :genCode="mcpValidator.genVerifyCode"
+    :clearCode="mcpValidator.clearVerifyCode" v-model:show="showChat" />
   <div class="ai-chat-toggle" title="打开AI对话框" v-show="!showChat" @click="showChat = true">
     <img :src="AiChatSvg" />
   </div>
