+++ demo/src/components/ChatDialog.vue
@@ -16,279 +16,64 @@
         @cancel="abortRequest"></tr-sender>
     </template>
   </tr-container>
-  <div @click="
+  <div @click="() => {
     show = !show;
-  senderRef?.focus();
-  ">
+    senderRef?.focus();
+  }">
     <slot></slot>
   </div>
 </template>
 
 <script setup>
-import { IconAi, IconUser } from '@opentiny/tiny-robot-svgs';
-import { h, ref, watch, nextTick, computed, onUnmounted } from 'vue';
-import { BaseModelProvider, AIClient, useMessage, STATUS, GeneratingStatus } from '@opentiny/tiny-robot-kit';
+import { AIClient, useMessage, STATUS, GeneratingStatus} from '@opentiny/tiny-robot-kit';
+import { ref, watch, nextTick, computed, onUnmounted } from 'vue';
+import { CustomModelProvider } from './chat-config/custom-model-provider'
+import { SimpleToolCallHandler } from './chat-config/simple-tool-call-handler';
+import { roles, promptItems } from './chat-config/chat-config';
 
 const props = defineProps({
-  clientId: { type: String, default: () => '' },
+  getClientId: { type: String, default: () => () => '' },
   genCode: { type: Function, default: () => () => { } },
   clearCode: { type: Function, default: () => () => { } },
   memory: { type: Boolean, default: true },
 });
 
-class SimpleToolCallHandler {
-  constructor() {
-    this.styleElementIds = new Set();
-    this.initStyles();
-    this.updateToolTimer = null;
+const show = defineModel('show', {
+  type: Boolean,
+  default: true,
+});
+
+const fullscreen = ref(false);
+const senderRef = ref(null);
+
+// 配置AI对话提供商
+const customModelProvider = new CustomModelProvider(
+  {
+    memory: props.memory,
+  }, 
+  {
+    toolCallHandler: new SimpleToolCallHandler(),
+    validator: {
+      genCode: props.genCode,
+      clearCode: props.clearCode,
+    },
+    getClientId: props.getClientId
   }
-
-  handler(extra, handler) {
-    const element = this.getElement(extra);
-    if (!element) {
-      const { onData } = handler;
-      onData({ choices: [{ delta: { content: this.createElement(extra) } }] });
-      this.updateToolTimer = setTimeout(() => {
-        this.updateTool(extra);
-        this.updateToolTimer = null;
-      }, 0);
-    } else {
-      if (this.updateToolTimer) {
-        clearTimeout(this.updateToolTimer);
-        this.updateToolTimer = null;
-      }
-      this.updateTool(extra);
-    }
-  }
-
-  createElement(extra) {
-    this.createStyle(extra);
-    return `<div class="tool-call" id="${extra.toolCall.id}"></div>`;
-  }
-
-  createStyle(extra) {
-    const style = document.createElement('style');
-    style.id = `tool_call_${extra.toolCall.id}`;
-    document.head.appendChild(style);
-    this.styleElementIds.add(style.id);
-  }
-
-  getStyle(extra) {
-    return document.querySelector(`#tool_call_${extra.toolCall.id}`);
-  }
-
-  getElement(extra) {
-    return document.querySelector(`div.tool-call#${extra.toolCall.id}`)
-  }
-
-  updateTool(extra) {
-    const element = this.getElement(extra);
-    const style = this.getStyle(extra);
-    if (!element || !style) {
-      console.warn('no tool call info')
-      return;
-    }
-
-    if (extra.callToolResult) {
-      style.innerHTML = `
-       .tool-call#${extra.toolCall.id}::after {
-         content: '调用工具 ${extra.toolCall.function.name} ${extra.callToolResult.isError ? '失败 ❌' : '成功 ✅'}'
-       }
-      `
-    } else {
-      style.innerHTML = `
-       .tool-call#${extra.toolCall.id}::after {
-         content: '正在调用工具 ${extra.toolCall.function.name} ...'
-       }
-         `
-    }
-  }
-  initStyles() {
-    const style = document.createElement('style');
-    style.id = 'simple-tool-call-handler-base-styles';
-    style.innerHTML = `
-      .chat-dialog div.tool-call {
-        padding: 8px 16px;
-        margin: 12px 0;
-        background: #EFEFEF;
-        border: #EEE 1px solid;
-        border-radius: 10px;
-      }
-    `
-    document.head.appendChild(style);
-    this.styleElementIds.add(style.id);
-  }
-
-  cleanup() {
-    if (this.updateToolTimer) {
-      clearTimeout(this.updateToolTimer);
-      this.updateToolTimer = null;
-    }
-
-    this.styleElements.forEach(styleId => {
-      const element = document.getElementById(styleId);
-      element?.remove();
-    });
-    this.styleElements.clear();
-  }
-}
-
-// 自定义模型提供者
-class CustomModelProvider extends BaseModelProvider {
-  toolCallHandler = new SimpleToolCallHandler();
-  constructor(options) {
-    super(options);
-  }
-  validateRequest() { }
-  async getData(request) {
-    this.validateRequest(request);
-
-    const verifyCode = await props.genCode();
-    const lastMessage = request.messages[request.messages.length - 1].content;
-    const options = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'connector-client-id': props.clientId,
-        'mcp-verify-code': verifyCode,
-      },
-      body: JSON.stringify(props.memory ? { messages: request.messages } : { query: lastMessage }),
-    };
-
-    const response = await fetch(`http://localhost:3001/chat`, options);
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      throw new Error(`HTTP error! status: ${response.status}, details: ${errorText}`);
-    }
-    return response;
-  }
-
-  async chat(request) {
-    try {
-      const response = await this.getData(request);
-      const reader = response.body.getReader();
-      const decoder = new TextDecoder('utf-8');
-      let text = '';
-
-      // 逐块读取流数据
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) {
-          console.log('Stream complete');
-          break;
-        }
-        const chunk = decoder.decode(value, { stream: true });
-
-        try {
-          const message = JSON.parse(chunk.slice(6));
-          console.log(message); // 输出流的每一部分
-          text += message.choices[0].delta.content;
-        } catch (error) {
-          text += '';
-        }
-      }
-
-      return { choices: [{ message: { content: text } }] };
-    } catch (error) {
-      console.error(error);
-      throw error;
-    } finally {
-      props.clearCode();
-    }
-  }
-
-  async chatStream(request, handler) {
-    const { onData, onDone, onError } = handler;
-    let reader = null;
-    try {
-      const response = await this.getData(request);
-      reader = response.body.getReader();
-      const decoder = new TextDecoder('utf-8');
-      let buffer = '';
-
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-        // Append new chunk to buffer
-        buffer += decoder.decode(value, { stream: true });
-        // Process complete lines from buffer
-        while (true) {
-          const lineEnd = buffer.indexOf('\n');
-          if (lineEnd === -1) break;
-          const line = buffer.slice(0, lineEnd).trim();
-          buffer = buffer.slice(lineEnd + 1);
-          if (line.startsWith('data: ')) {
-            const data = line.slice(6);
-            if (data === '[DONE]') break;
-            try {
-              const parsed = JSON.parse(data);
-              if (parsed.choices[0].delta.extra?.toolCall) {
-                const extra = parsed.choices[0].delta.extra;
-                this.toolCallHandler.handler(extra, handler);
-                continue;
-              }
-              const content = parsed.choices[0].delta.content;
-              if (content) {
-                onData({ choices: [{ delta: { content } }] });
-              }
-            } catch (e) {
-              // Ignore invalid JSON
-            }
-          }
-        }
-      }
-      onDone();
-    } catch (error) {
-      onError(error);
-      throw error;
-    } finally {
-      reader.cancel();
-      props.clearCode();
-    }
-  }
-
-  destroy() {
-    this.toolCallHandler.cleanup();
-  }
-
-}
-
-const customModelProvider = new CustomModelProvider();
+);
 onUnmounted(() => { customModelProvider.destroy(); });
 
+// 配置AI对话客户端
 const client = new AIClient({
   provider: 'custom',
   providerImplementation: customModelProvider,
 });
 
-// 使用tiny-robot 提供的API
 const { messages, inputMessage, messageState, sendMessage, abortRequest } = useMessage({
   client,
   useStreamByDefault: true,
   initialMessages: [],
 });
 
-const promptItems = [
-  {
-    label: '列出工具',
-    description: '列出目前系统中可用的工具！',
-    icon: h('span', { style: { fontSize: '18px' } }, '🧠'),
-    badge: 'NEW',
-  },
-  {
-    label: '界面操作',
-    description: '通过界面新增用户 张三 男 2000-1-1',
-    icon: h('span', { style: { fontSize: '18px' } }, '🧠'),
-    badge: 'NEW',
-  },
-  {
-    label: '函数调用',
-    description: '新增用户 李四 女 2000-2-2',
-    icon: h('span', { style: { fontSize: '18px' } }, '🧠'),
-    badge: 'NEW',
-  },
-];
 
 const handlePromptItemClick = (e, item) => {
   sendMessage(item.description);
@@ -307,33 +92,6 @@ const showMessages = computed(() => {
   }
   return messages.value;
 });
-const show = defineModel('show', {
-  type: Boolean,
-  default: true,
-});
-const fullscreen = ref(false);
-const senderRef = ref(null);
-
-const aiAvatar = h(IconAi, { style: { fontSize: '32px' } });
-const userAvatar = h(IconUser, { style: { fontSize: '32px' } });
-
-// 定义角色图标以及样式
-const roles = {
-  assistant: {
-    placement: 'start',
-    avatar: aiAvatar,
-    maxWidth: '90%',
-    type: 'markdown',
-    mdConfig: { html: true },
-  },
-  user: {
-    placement: 'end',
-    avatar: userAvatar,
-    maxWidth: '90%',
-    type: 'markdown',
-    mdConfig: { html: true },
-  },
-};
 
 // 最新消息滚动到底部
 watch(
